claude(scan-git-for-tils): break up a couple modules
