+++ tools/claude/config/skills/scan-git-for-tils/op/__init__.py
@@ -1,0 +1,3 @@
+"""1Password integration utilities."""
+
+from __future__ import annotations
